+++ crates/mimic/src/design/base/types/time.rs
@@ -1,4 +1,4 @@
-use crate::design::{base::types, prelude::*};
+use crate::design::prelude::*;
 
 ///
 /// Milliseconds
@@ -20,33 +20,6 @@ pub struct Seconds {}
 
 #[newtype(primitive = "Nat64", item(prim = "Nat64"))]
 pub struct Minutes {}
-
-///
-/// Timestamp
-///
-
-#[newtype(primitive = "Nat64", item(prim = "Nat64"))]
-pub struct Timestamp {}
-
-impl Timestamp {
-    #[must_use]
-    pub fn now() -> Self {
-        Self(icu::utils::time::now_secs())
-    }
-}
-
-///
-/// Created
-///
-/// A Timestamp that defaults to the current now() time when it is created
-///
-
-#[newtype(
-    primitive = "Nat64",
-    item(is = "types::time::Timestamp"),
-    default = "types::time::Timestamp::now"
-)]
-pub struct Created {}
 
 ///
 /// Duration